[FEAT] 플레이어 턴 공격 구현, 테스트 필요

+++ backend/gameproject/src/main/java/com/example/gameproject/db/repository/MyCharacterRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface MyCharacterRepository extends JpaRepository<MyCharacter, Long> {
 
     @Query(value = "select * from my_character c where c.user_id = :userId ORDER BY c.pos ASC", nativeQuery = true)
-    List<MyCharacter> getMyCharacters(@Param("userId") int userId);
+    List<MyCharacter> getMyCharacters(@Param("userId") Long userId);
 
     @Query("select mc from MyCharacter mc where mc.user.id=:id")
     List<MyCharacter> findByUserId(@Param("id") long userId);
